requirements.txt
openai
supabase
python-dotenv
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6